requirement.txt
Flask==2.0.3
mysql-connector-python==8.0.26
pdf2docx==0.5.1
docx2pdf==0.1.7
comtypes==1.1.9
PyPDF2==3.0.0
pandas==1.3.3
openpyxl==3.0.9
